Rework the image plugin
This commit makes a lot of changes around the image plugin:
* Introduces the concept of image_plugin and privileged_image_plugin.
When an unprivileged/privileged container has volume related operations
(create/destroy/metrics), the correct plugin should be called with the
provided global arguments (image_plugin_extra_args and
privileged_plugin_extra_args respectively). They should also be called
as the garden user or unprivileged user respectively.

* Calls the right plugin on metrics and both plugins on destroy, since
We can now do the right thing and call the right plugin (or both),
as the right user, passing only the handle. We no longer have to pass the
rootfs path to a plugin to determine which store to use (this was a hack for
GrootFS's sake originally).

* Removes calling the `clean` command on the image plugin as we have
decided the plugin should be in charge of when it performs cleanup.

This commit also makes significant changes from an engineering view:
* From an engineering point of view, we have split the creation of
plugin commands from their execution, making smaller components with
easier testing.

* The top level integration tests now make use of the `extra_args`
garden arguments to pass test information to the fake plugin. This
increases clarity in the test both because expected values are not
hidden in a binary and because the fake plugin could be rewritten with a
real CLI library.

* Determining whether a container was privileged during metrics
requests is now made by loading the bundle json from disk and checking
whether the user namespace exists in the set of configured namespaces.
This is unfortunately slightly coupled to our meaning of "privileged"
and also may have some performance degradation during metrics calls -
which now have to touch disk. We could improve this by storing a
property in the PropertyManager if it becomes an issue.

* Destroy calls both plugins now, so image plugins must be idempotent,
like the rest of the Garden destroy stack.

[#136343839 #137888407 #137888425]